refactor: make map update
