+++ frontend/src/components/MainView.tsx
@@ -2,14 +2,13 @@ import { AnimatePresence } from "motion/react";
 import { useEffect, useState } from "react";
 import Exactum2 from "../assets/exactum-2.min.svg?react";
 import { useMapTransform } from "../hooks/useMapTransform";
-import { findAllRooms, findRoomById } from "../services/roomsService";
+import { useRoomProperties } from "../hooks/useRoomProperties";
+import { findRoomById } from "../services/roomsService";
 import type { Room } from "../types";
 import ColorToggle from "./ColorToggle";
 import "./MainView.css";
 import SidePanel from "./SidePanel";
 import ZoomButtons from "./ZoomButtons";
-
-const ROOM_LABEL_FONT_SIZE = 24;
 
 function MainView() {
   const {
@@ -19,98 +18,12 @@ function MainView() {
     handleZoomFunc,
     handleResetFunc,
   } = useMapTransform();
+
+  const { useAvailability, setUseAvailability, onUpdate } = useRoomProperties();
+
   const [activeRoomId, setActiveRoomId] = useState<string | null>(null);
   const [isSidePanelOpen, setIsSidePanelOpen] = useState<boolean>(false);
   const [room, setRoom] = useState<Room | null>(null);
-
-  function createRoomInfoLabel(
-    centerX: number,
-    centerY: number,
-    lines: string[],
-  ): SVGTextElement {
-    const text = document.createElementNS("http://www.w3.org/2000/svg", "text");
-    text.setAttribute("x", String(centerX));
-    text.setAttribute("y", String(centerY));
-    text.setAttribute("text-anchor", "middle");
-    text.setAttribute("dominant-baseline", "middle");
-    text.classList.add("room-label");
-
-    const fontSize = ROOM_LABEL_FONT_SIZE;
-    const lineHeight = fontSize * 1.2;
-    // Offset the starting position so the label is centered in the middle of the room
-    const offsetStart = -((lines.length - 1) / 2) * lineHeight;
-
-    lines.forEach((line, i) => {
-      const tspan = document.createElementNS(
-        "http://www.w3.org/2000/svg",
-        "tspan",
-      );
-      tspan.setAttribute("x", String(centerX));
-      tspan.setAttribute(
-        "dy",
-        i === 0 ? String(offsetStart) : String(lineHeight),
-      );
-      tspan.textContent = line;
-      text.appendChild(tspan);
-    });
-
-    return text;
-  }
-
-  useEffect(() => {
-    async function mapDataToRoomElements() {
-      try {
-        const result = await findAllRooms();
-
-        const roomsMap = new Map(result.map((room) => [room.name, room]));
-
-        const roomElements = document.querySelectorAll("path[data-room]");
-
-        roomElements.forEach((element) => {
-          const roomName = element.getAttribute("data-room");
-          if (roomName && roomsMap.has(roomName)) {
-            const room = roomsMap.get(roomName);
-
-            if (room) {
-              element.id = String(room.id);
-              element.classList.add("room");
-
-              if (element instanceof SVGGraphicsElement) {
-                const bbox = element.getBBox();
-                const centerX = bbox.x + bbox.width / 2;
-                const centerY = bbox.y + bbox.height / 2;
-
-                const lines = [
-                  roomName,
-                  `${room.area}m²`,
-                  `${room.contracts.length}/${room.capacity}`,
-                ];
-
-                const group = document.createElementNS(
-                  "http://www.w3.org/2000/svg",
-                  "g",
-                );
-                group.classList.add("room-group");
-
-                const label = createRoomInfoLabel(centerX, centerY, lines);
-
-                element.parentNode?.insertBefore(group, element);
-                group.appendChild(element);
-                group.appendChild(label);
-              }
-            }
-          }
-        });
-      } catch (error: unknown) {
-        let errorMessage = "❌ Failed to map rooms: ";
-        if (error instanceof Error) {
-          errorMessage += error.message;
-        }
-        console.log(errorMessage);
-      }
-    }
-    mapDataToRoomElements();
-  }, []);
 
   useEffect(() => {
     const rooms = document.querySelectorAll("path[data-room]");
@@ -147,6 +60,11 @@ function MainView() {
     }
   }
 
+  async function onRoomChange() {
+    findRoom(activeRoomId!);
+    onUpdate();
+  }
+
   return (
     <>
       <div className="wrapper">
@@ -161,7 +79,10 @@ function MainView() {
           handleReset={handleResetFunc}
         />
 
-        <ColorToggle />
+        <ColorToggle
+          useAvailability={useAvailability}
+          setUseAvailability={setUseAvailability}
+        />
 
         <AnimatePresence>
           {isSidePanelOpen && (
@@ -171,7 +92,7 @@ function MainView() {
                 setIsSidePanelOpen(false);
                 setActiveRoomId(null);
               }}
-              onPersonSaved={() => findRoom(activeRoomId!)}
+              onPersonSaved={() => onRoomChange()}
             />
           )}
         </AnimatePresence>
